backend-api/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pydantic
python-multipart
mysql-connector-python
pymysql

